device-api-e2e/requirements.txt
allure-behave==2.9.45
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==22.1.0
behave==1.2.6
# black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
curlify==2.2.1
# flake8==4.0.1
greenlet==1.1.2
grpcio==1.47.0
grpcio-tools==1.47.0
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.20.1
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyHamcrest==2.0.3
pyparsing==3.0.9
pytest==7.1.2
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.11
